// src/types.ts
export interface AICarState {
  id: string;
  name: string;
  driverName?: string; // Tên nhân vật / tài xế lái xe
  color: string;
  hexColor: number;
  type: 'hypercar' | 'muscle' | 'formula' | 'gt_racer' | 'cyber_coupe';
  speed: number;
  targetSpeed: number;
  maxSpeed: number;
  acceleration: number;
  lap: number;
  lapProgress: number; // 0 to 1 along track
  lateralOffset: number; // offset from track center line (-1.0 to 1.0)
  targetLateralOffset: number;
  steerAngle: number;
  rank: number;
  aggression: number;
  isDrifting: boolean;
  driftAngle: number;
  collisionCooldown: number;
  meshIndex: number;
  inTunnel?: boolean;
}